requirements.txt
pytest~=7.4.0
--find-links https://download.pytorch.org/whl/torch_stable.html
torch>=2.1.2+cu118
torchvision>=0.16.2+cu118
opencv-python~=4.8.1.78
numpy~=1.24.3
pillow~=9.4.0
setuptools~=68.0.0
tensorflow~=2.15.0
six~=1.16.0
lxml~=4.9.3
tqdm~=4.65.0